Move log segment into separate module
Log segment is going to be used by both snapshot and table changes.
It makes sense to separate it into its own module
