+++ kernel/src/lib.rs
@@ -69,6 +69,7 @@ pub mod path;
 #[cfg(not(feature = "developer-visibility"))]
 pub(crate) mod path;
 
+pub mod log_segment;
 pub mod scan;
 pub mod schema;
 pub mod snapshot;
